refactor: change structure from v1 to v2
