+++ src/v2/commands/shuffle.ts
